+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Avoid checking trust access to the VBA project object model twice on function generation.
 - Return raw result when using as a python module even if HTTP 202 is received.
 - Return the full response as text when using as a python module, and not only the first 255 characters.
+- In case NO_PROXY environment variable was containing multiple URLs and should have been applied, it was not the case by the add-in.
 
 ### Removed
 - Drop support for `python` < `3.8`.
@@ -107,6 +108,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `caching` options are now per formula instead of per REST API and have been renamed. See [migration guide](docs/migration_guide.md) for more details.
 - xlwings configuration is now retrieved faster (not retrieved from a file and do not look for environment variable).
 - `definition_retrieval_auths` parameters have been renamed to match `requests_auth` parameters for consistency with `auth` section.
+- Update is now retrieved as soon as possible instead of only at Microsoft Excel close.
 
 ## [0.69.0] - 2018-12-03
 ### Changed